[#1052] Downloading data fails on azure k8s
Model training with data did not pass the EDI validation because we
forgot to expand the map variable with correct connection types.

To prevent it in the future, we added robot tests for different
combinations of connection URI, localPath, and remotePath parameters.

Additionally, the parsing of azure URL has been fixed in the function,
which creates RClone config.