[#1037] Remove sensitive information from feedback
Some headers from a model request can contain some sensitive
information: token, password, and so on. The feedback collector service
must not send it to Fluentd service.

To fix it, we introduce the new config option that allows specifying
a list of forbidden headers. A user can pass this parameter through
helm values too.
The default list of forbidden headers is:
  - authorization
  - x-jwt
  - x-user
  - x-email

Moreover, we improve robot feedback test to check that feedback does
not contain forbidden headers.